requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=8.3.1
pyyaml>=5.4.1
tqdm>=4.62.2
matplotlib>=3.4.3
scikit-learn>=0.24.2
fastapi
uvicorn[standard]